Move xyz_to_list to own file resultlist.py
