+++ ospd/ospd.py
@@ -27,7 +27,6 @@ import multiprocessing
 import time
 import os
 
-from collections import OrderedDict
 from typing import (
     List,
     Any,
@@ -1317,32 +1316,6 @@ class OSPDaemon:
             qod,
         )
 
-    def add_scan_log_to_list(
-        self,
-        result_list: List,
-        host: str = '',
-        hostname: str = '',
-        name: str = '',
-        value: str = '',
-        port: str = '',
-        test_id: str = '',
-        qod: str = '',
-    ) -> List[Dict[str, str]]:
-        """ Adds log result to a list of results. """
-        result_list = self.add_result_to_list(
-            result_list,
-            ResultType.LOG,
-            host,
-            hostname,
-            name,
-            value,
-            port,
-            test_id,
-            '0.0',
-            qod,
-        )
-        return result_list
-
     def add_scan_error(
         self,
         scan_id: str,
@@ -1365,29 +1338,6 @@ class OSPDaemon:
             test_id,
         )
 
-    def add_scan_error_to_list(
-        self,
-        result_list: Optional[List],
-        host: str = '',
-        hostname: str = '',
-        name: str = '',
-        value: str = '',
-        port: str = '',
-        test_id='',
-    ) -> List[Dict[str, str]]:
-        """ Adds an error result to result list. """
-        result_list = self.add_result_to_list(
-            result_list,
-            ResultType.ERROR,
-            host,
-            hostname,
-            name,
-            value,
-            port,
-            test_id,
-        )
-        return result_list
-
     def add_scan_host_detail(
         self,
         scan_id: str,
@@ -1400,20 +1350,6 @@ class OSPDaemon:
         self.scan_collection.add_result(
             scan_id, ResultType.HOST_DETAIL, host, hostname, name, value
         )
-
-    def add_scan_host_detail_to_list(
-        self,
-        result_list: Optional[List] = None,
-        host: str = '',
-        hostname: str = '',
-        name: str = '',
-        value: str = '',
-    ) -> List[Dict[str, str]]:
-        """ Adds a host detail result to result list. """
-        result_list = self.add_result_to_list(
-            result_list, ResultType.HOST_DETAIL, host, hostname, name, value,
-        )
-        return result_list
 
     def add_scan_alarm(
         self,
@@ -1440,61 +1376,3 @@ class OSPDaemon:
             severity,
             qod,
         )
-
-    def add_scan_alarm_to_list(
-        self,
-        result_list: Optional[List],
-        host: str = '',
-        hostname: str = '',
-        name: str = '',
-        value: str = '',
-        port: str = '',
-        test_id: str = '',
-        severity: str = '',
-        qod: str = '',
-    ) -> List[Dict[str, str]]:
-        """ Adds an alarm result to a result list. """
-        result_list = self.add_result_to_list(
-            result_list,
-            ResultType.ALARM,
-            host,
-            hostname,
-            name,
-            value,
-            port,
-            test_id,
-            severity,
-            qod,
-        )
-        return result_list
-
-    def add_result_to_list(
-        self,
-        results: List[Dict[str, str]],
-        result_type: int,
-        host: str = '',
-        hostname: str = '',
-        name: str = '',
-        value: str = '',
-        port: str = '',
-        test_id: str = '',
-        severity: str = '',
-        qod: str = '',
-    ) -> List[Dict[str, str]]:
-
-        result = OrderedDict()  # type: Dict
-        result['type'] = result_type
-        result['name'] = name
-        result['severity'] = severity
-        result['test_id'] = test_id
-        result['value'] = value
-        result['host'] = host
-        result['hostname'] = hostname
-        result['port'] = port
-        result['qod'] = qod
-
-        if results is None:
-            results = list()
-        results.append(result)
-
-        return results
